+++ src/controllers/IngredientsController.js
@@ -6,7 +6,16 @@ class IngredientsController {
     const { id } = request.params;
     const { name } = request.body;
 
+    if (!id) {
+      throw new AppError("ID do ingrediente não fornecido.");
+    }
+
+    if (!name) {
+      throw new AppError("O novo nome do ingrediente não foi fornecido.");
+    }
+
     const ingredient = await knex("ingredients").where({ id }).first();
+
     if (!ingredient) {
       throw new AppError("Ingrediente não encontrado!");
     }
@@ -23,7 +32,15 @@ class IngredientsController {
   async delete(request, response) {
     const { id } = request.params;
 
+    if (!id) {
+      throw new AppError("ID do ingrediente não fornecido.");
+    }
+
     await knex("ingredients").where({ id }).delete();
+
+    if (!ingredient) {
+      throw new AppError("Ingrediente não encontrado!");
+    }
 
     return response.json();
   }
